+++ components/entries/entry.js
@@ -1,9 +1,9 @@
-import { deleteEntryToServer } from "../../../../api.js";
-import { sharedState } from "../../../../store/state.js";
-import { renderCategoryOptions } from "../../../../components/inputForm/inputFormItems/categoryRender.js";
-import { updateTotalAmounts } from "../../../../components/totalAmount/totalAmount-util.js";
-import { createModal } from "../../../../components/modal.js";
-import { store } from "../../../../store/store.js";
+import { deleteEntryToServer } from "../../api.js";
+import { sharedState } from "../../store/state.js";
+import { renderCategoryOptions } from "../inputForm/inputFormItems/categoryRender.js";
+import { updateTotalAmounts } from "../totalAmount/totalAmount-util.js";
+import { createModal } from "../modal.js";
+import { store } from "../../store/store.js";
 import { updateDateSectionTotals } from "./entry-util.js";
 
 export function deleteEntry() {
@@ -45,7 +45,7 @@ export function deleteEntryConfirm(deleteBtn) {
 
   updateTotalAmounts();
 
-  import("../../calendar/calendarTotalAmount.js").then((module) => {
+  import("../../pages/main/calendar/calendarTotalAmount.js").then((module) => {
     module.updateCalendarTotalAmount();
   });
 }
