+++ components/header/dateRender.js
@@ -1,6 +1,6 @@
 import { sharedState } from "../../../store/state.js";
 import { loadEntriesFromServer } from "../../../api.js";
-import { renderOneEntiry } from "../../pages/main/ledger/entries/entry-util.js";
+import { renderOneEntiry } from "../entries/entry-util.js";
 import { store } from "../../../store/store.js";
 import { updateTotalAmounts } from "../totalAmount/totalAmount-util.js";
 import { initCalendarView } from "../../pages/main/calendar/calendarView.js";
